exposes leaflet map onEachFeature content to app;
